Moved non-routine-specific API functions and includes to separate files
